+++ 0x15-file_io/main.h
@@ -12,6 +12,6 @@ ssize_t read_textfile(const char *filename, size_t letters);
 int create_file(const char *filename, char *text_content);
 int append_text_to_file(const char *filename, char *text_content);
 char *_buffer(char *file, int bytes_selected);
-void _close(ssize_t origin, ssize_t destiny);
+void _close(int origin, int destiny);
 
 #endif /* _MAIN_H_ */
